Public Documentation

Documentation for Imaginocean.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Imaginocean

Imaginocean.heatsphere! — Function

heatsphere!(axis::Axis3, field::Field, k_index=1; kwargs...)

A heatmap of an Oceananigans.jl Field on the sphere at vertical index k_index.

Arguments

axis :: Makie.Axis3: a 3D axis.

field :: Oceananigans.Field: an Oceananigans.jl field with non-flat horizontal dimensions.

k_index :: Int: The integer corresponding to the vertical index of the field to visualize; default: 1.

Accepts all keyword arguments for Makie.mesh! method.